allow key/value pairs for enum fields
Allowing for objects to be passed for enum fields gives us the option to
set option text and values separately. This commit allows form authors
to use arrays for simple enum fields but also extends the syntax so that
if an object is passed, the keys of the object are used for option
values and the values of the object are used for option text.